Convert all plugins to classes
This helps with correctness while still exposing constructors. Now there should be no distinction between onEnable and construction of the controller.
