+++ src/globals/window.ts
@@ -8,7 +8,7 @@ import {
   SegmentedControlComponent,
   TooltipComponent,
 } from "../components/desmosComponents";
-import { GenericSettings, TransparentPlugins } from "../plugins";
+import { GenericSettings, PluginID, TransparentPlugins } from "../plugins";
 import CalcType from "./Calc";
 import { ItemModel } from "./models";
 
@@ -17,9 +17,9 @@ interface windowConfig extends Window {
   DesModder: any;
   DSM: TransparentPlugins;
   DesModderPreload?: {
-    pluginsForceDisabled: Set<string>;
-    pluginsEnabled: Record<string, boolean>;
-    pluginSettings: Record<string, GenericSettings>;
+    pluginsForceDisabled: Set<PluginID>;
+    pluginsEnabled: Record<PluginID, boolean | undefined>;
+    pluginSettings: Record<PluginID, GenericSettings | undefined>;
   };
 }
