+++ src/plugins/better-evaluation-view/index.ts
@@ -1,23 +1,11 @@
+import { PluginController } from "../PluginController";
 import "./better-evaluation-view.less";
 import { configList } from "./config";
-import { Calc } from "globals/window";
 import { Plugin } from "plugins";
 
-const betterEvaluationView: Plugin = {
-  id: "better-evaluation-view",
-  onEnable: () => {
-    Calc.controller.updateViews();
-    return undefined;
-  },
-  onDisable: () => {},
-  afterDisable: () => {
-    Calc.controller.updateViews();
-  },
-  config: configList,
-  onConfigChange: () => {
-    Calc.controller.updateViews();
-  },
-  enabledByDefault: true,
-  /* Has module overrides */
-};
-export default betterEvaluationView;
+export default class BetterEvaluationView extends PluginController {
+  static id = "better-evaluation-view" as const;
+  static enabledByDefault = true;
+  static config = configList;
+}
+BetterEvaluationView satisfies Plugin;
