+++ src/plugins/builtin-settings/index.ts
@@ -1,12 +1,10 @@
+import { PluginController } from "../PluginController";
 import { Config, configList } from "./config";
 import { Calc } from "globals/window";
-import MainController from "main/Controller";
 import { Plugin } from "plugins";
 import { getQueryParams } from "utils/depUtils";
 
 const managedKeys = configList.map((e) => e.key);
-
-let initialSettings: null | Config = null;
 
 function updateSettings(config: Config) {
   let { graphpaper, zoomButtons } = config;
@@ -18,46 +16,42 @@ function updateSettings(config: Config) {
   Calc.updateSettings({ ...config, zoomButtons, graphpaper });
 }
 
-function onEnable(_controller: MainController, config: Config) {
-  initialSettings = { ...config };
-  const queryParams = getQueryParams();
-  for (const key of managedKeys) {
-    initialSettings[key] =
-      (
-        Calc.settings as typeof Calc.settings & {
-          advancedStyling: boolean;
-          authorFeatures: boolean;
-        }
-      )[key] ?? false;
-  }
-  const queryConfig: Partial<Config> = {};
-  for (const key of managedKeys) {
-    if (queryParams[key]) {
-      queryConfig[key] = true;
-    }
-    if (queryParams["no" + key]) {
-      queryConfig[key] = false;
-    }
-  }
-  updateSettings(config);
-  return undefined;
-}
+export default class BuiltinSettings extends PluginController {
+  static id = "builtin-settings" as const;
+  static enabledByDefault = true;
+  static config = configList;
+  initialSettings: null | Config = null;
 
-function onDisable() {
-  if (initialSettings !== null) {
-    updateSettings(initialSettings);
-  }
-}
-
-const builtinSettings: Plugin = {
-  id: "builtin-settings",
-  onEnable,
-  onDisable,
-  enabledByDefault: true,
-  config: configList,
-  onConfigChange(config: Config) {
-    // called only when plugin is active
+  afterEnable(config: Config) {
+    this.initialSettings = { ...config };
+    const queryParams = getQueryParams();
+    for (const key of managedKeys) {
+      this.initialSettings[key] =
+        (
+          Calc.settings as typeof Calc.settings & {
+            advancedStyling: boolean;
+            authorFeatures: boolean;
+          }
+        )[key] ?? false;
+    }
+    const queryConfig: Partial<Config> = {};
+    for (const key of managedKeys) {
+      if (queryParams[key]) {
+        queryConfig[key] = true;
+      }
+      if (queryParams["no" + key]) {
+        queryConfig[key] = false;
+      }
+    }
     updateSettings(config);
-  },
-};
-export default builtinSettings;
+  }
+
+  afterDisable() {
+    if (this.initialSettings !== null) updateSettings(this.initialSettings);
+  }
+
+  onConfigChange(config: Config) {
+    updateSettings(config);
+  }
+}
+BuiltinSettings satisfies Plugin;
