+++ src/plugins/find-replace/index.ts
@@ -1,3 +1,4 @@
+import { PluginController } from "../PluginController";
 import Controller from "./Controller";
 import View from "./View";
 import { Calc, Console } from "globals/window";
@@ -19,30 +20,27 @@ function tryInitView() {
   }
 }
 
-function onEnable() {
-  if (Calc.controller.getExpressionSearchOpen()) {
-    tryInitView();
-  }
-  dispatchListenerID = Calc.controller.dispatcher.register(({ type }) => {
-    if (type === "open-expression-search") {
+export default class FindReplace extends PluginController {
+  static id = "find-and-replace" as const;
+  static enabledByDefault = true;
+
+  afterEnable() {
+    if (Calc.controller.getExpressionSearchOpen()) {
       tryInitView();
-    } else if (type === "close-expression-search") {
-      view.destroyView();
     }
-    // may want to listen to update-expression-search-str
-  });
-  return controller;
-}
+    dispatchListenerID = Calc.controller.dispatcher.register(({ type }) => {
+      if (type === "open-expression-search") {
+        tryInitView();
+      } else if (type === "close-expression-search") {
+        view.destroyView();
+      }
+      // may want to listen to update-expression-search-str
+    });
+  }
 
-function onDisable() {
-  Calc.controller.dispatcher.unregister(dispatchListenerID);
-  view.destroyView();
+  afterDisable() {
+    Calc.controller.dispatcher.unregister(dispatchListenerID);
+    view.destroyView();
+  }
 }
-
-const findAndReplace: Plugin = {
-  id: "find-and-replace",
-  onEnable,
-  onDisable,
-  enabledByDefault: true,
-};
-export default findAndReplace;
+FindReplace satisfies Plugin;
