+++ src/plugins/performance-info/PerformanceView.tsx
@@ -1,4 +1,4 @@
-import Controller from "./Controller";
+import PerformanceInfo from ".";
 import "./PerformanceView.less";
 import { Component, jsx } from "DCGView";
 import { Button, IconButton, Tooltip } from "components";
@@ -6,7 +6,7 @@ import { format } from "i18n/i18n-core";
 import MainController from "main/Controller";
 
 export class PerformanceView extends Component<{
-  controller: () => Controller;
+  controller: () => PerformanceInfo;
   mainController: () => MainController;
 }> {
   template() {
@@ -89,7 +89,7 @@ export class PerformanceView extends Component<{
 }
 
 export function MainPopupFunc(
-  performanceViewController: Controller,
+  performanceViewController: PerformanceInfo,
   mainController: MainController
 ): PerformanceView {
   return (
