+++ src/plugins/pin-expressions/index.ts
@@ -3,7 +3,10 @@ import "./pinExpressions.less";
 import { Calc } from "globals/window";
 import { Plugin } from "plugins";
 
-export class PinExpressions extends PluginController {
+export default class PinExpressions extends PluginController {
+  static id = "pin-expressions" as const;
+  static enabledByDefault = false;
+
   pinExpression(id: string) {
     if (Calc.controller.getItemModel(id)?.type !== "folder")
       this.controller.updateExprMetadata(id, {
@@ -33,19 +36,4 @@ export class PinExpressions extends PluginController {
     el?.classList.toggle("dsm-has-pinned-expressions", hasPinnedExpressions);
   }
 }
-
-const pinExpressions: Plugin = {
-  id: "pin-expressions",
-  // Controller handles enable/disable by changing the results of isPinned
-  // (used in modified module definitions), but we need to update views
-  onEnable: (controller) => {
-    Calc.controller.updateViews();
-    return new PinExpressions(controller);
-  },
-  onDisable: () => {
-    Calc.controller.updateViews();
-  },
-  enabledByDefault: true,
-  /* Has module overrides */
-};
-export default pinExpressions;
+PinExpressions satisfies Plugin;
