+++ src/plugins/video-creator/backend/capture.ts
@@ -1,4 +1,4 @@
-import Controller from "../Controller";
+import VideoCreator from "..";
 import { scaleBoundsAboutCenter } from "./utils";
 import { Calc } from "globals/window";
 import MainController from "main/Controller";
@@ -9,17 +9,17 @@ let callbackIfCancel: (() => void) | null = null;
 
 export type CaptureMethod = "once" | "action" | "slider" | "ticks";
 
-export function cancelCapture(controller: Controller) {
+export function cancelCapture(controller: VideoCreator) {
   controller.captureCancelled = true;
   callbackIfCancel?.();
 }
 
-async function captureAndApplyFrame(controller: Controller) {
+async function captureAndApplyFrame(controller: VideoCreator) {
   const frame = await captureFrame(controller);
   controller.pushFrame(frame);
 }
 
-export async function captureFrame(controller: Controller) {
+export async function captureFrame(controller: VideoCreator) {
   const width = controller.getCaptureWidthNumber();
   const height = controller.getCaptureHeightNumber();
   const targetPixelRatio = controller.getTargetPixelRatio();
@@ -66,7 +66,7 @@ export interface SliderSettings {
   stepLatex: string;
 }
 
-export async function captureSlider(controller: Controller) {
+export async function captureSlider(controller: VideoCreator) {
   const sliderSettings = controller.sliderSettings;
   const variable = sliderSettings.variable;
   const min = EvaluateSingleExpression(sliderSettings.minLatex);
@@ -104,7 +104,7 @@ function slidersLatexJoined() {
     .join(";");
 }
 
-async function captureActionFrame(controller: Controller, step: () => void) {
+async function captureActionFrame(controller: VideoCreator, step: () => void) {
   let stepped = false;
   try {
     const tickCountRemaining = controller.getTickCountNumber();
@@ -140,7 +140,7 @@ async function captureActionFrame(controller: Controller, step: () => void) {
 }
 
 async function captureActionOrSliderTicks(
-  controller: Controller,
+  controller: VideoCreator,
   step: () => void
 ) {
   await new Promise<void>((resolve) => {
@@ -178,7 +178,7 @@ function forceReloadMenu(controller: MainController) {
   }
 }
 
-export async function capture(controller: Controller) {
+export async function capture(controller: VideoCreator) {
   controller.isCapturing = true;
   controller.updateView();
   const tickSliders = Calc.controller._tickSliders.bind(Calc.controller);
