+++ src/preload/moduleOverrides/show-tips.replacements
@@ -32,7 +32,7 @@ $DCGView.createElement(
     {
       predicate: () => DSM.showTips,
     },
-    () => DSM.showTips.createTipElement()
+    () => DSM.showTips?.createTipElement()
   )
 )
 ```
