+++ src/utils/messages.ts
@@ -32,9 +32,9 @@ type MessageWindowToContent =
 type MessageContentToWindow =
   | {
       type: "apply-initial-data";
-      pluginsEnabled: Record<PluginID, boolean>;
+      pluginsEnabled: Record<PluginID, boolean | undefined>;
       pluginsForceDisabled: PluginID[];
-      pluginSettings: Record<PluginID, GenericSettings>;
+      pluginSettings: Record<PluginID, GenericSettings | undefined>;
       scriptURL: string;
     }
   | HeartbeatError;
@@ -88,14 +88,14 @@ export function listenToMessageDown(
 
 /** Security issue on Firefox with posting a Map, so use this to convert a
  * Map to a Record (plain JS object). */
-export function mapToRecord<V>(x: Map<string, V>): Record<string, V> {
-  return Object.fromEntries(x.entries());
+export function mapToRecord<K extends string, V>(x: Map<K, V>): Record<K, V> {
+  return Object.fromEntries(x.entries()) as Record<K, V>;
 }
 
 /** Security issue on Firefox with posting a Map, so use this to convert a
  * Record (plain JS object) back to a Map. */
-export function recordToMap<V>(x: Record<string, V>): Map<string, V> {
-  return new Map(Object.entries(x));
+export function recordToMap<K extends string, V>(x: Record<K, V>): Map<K, V> {
+  return new Map(Object.entries(x)) as Map<K, V>;
 }
 
 /** Security issue on Firefox with posting a Set, so use this to convert a
